1. 별, 골드 받는 수량 변환 출력 ( 10.0A 등) 2. 아이템획득 연출 변경제제작 3. 풀링시스템 변경 4. 대규모 패치중 (너무많이해서 기억이잘안ㄴ남)

+++ Assets/CODE/Manager/DataManager.cs
@@ -47,10 +47,10 @@ public class DataManager : MonoBehaviour
         public int miniTicket;
 
         // 3. 버프 남은 시간
-        public double buffAtkTime;
-        public double buffGoldTime;
-        public double buffMoveSpeedTime;
-        public double buffBigAtkTime;
+        public int buffAtkTime;
+        public int buffGoldTime;
+        public int buffMoveSpeedTime;
+        public int buffBigAtkTime;
 
         // 4. 뉴비 혜택
         public int getNewbieRewardCount;
